refactor(credentials): extract keyring tests to sibling file
